+++ includes/admin/class-auto-load-next-post-admin-menus.php
@@ -3,6 +3,7 @@
  * Setup menus in the WordPress admin.
  *
  * @since    1.0.0
+ * @version  1.4.8
  * @author   Sébastien Dumont
  * @category Admin
  * @package  Auto Load Next Post
@@ -10,9 +11,8 @@
  */
 
 if ( ! defined('ABSPATH')) {
-	exit;
+	exit; // Exit if accessed directly.
 }
-// Exit if accessed directly
 
 if ( ! class_exists('Auto_Load_Next_Post_Admin_Menus')) {