+++ includes/auto-load-next-post-conditional-functions.php
@@ -1,20 +1,20 @@
 <?php
 /**
-	 * Auto Load Next Post Conditional Functions
-	 *
-	 * Functions for determining the current query/page.
-	 *
-	 * @since    1.0.0
-	 * @author   Sébastien Dumont
-	 * @category Core
-	 * @package  Auto Load Next Post
-	 * @license  GPL-2.0+
-	 */
+ * Auto Load Next Post Conditional Functions
+ *
+ * Functions for determining the current query/page.
+ *
+ * @since    1.0.0
+ * @version  1.4.8
+ * @author   Sébastien Dumont
+ * @category Core
+ * @package  Auto Load Next Post
+ * @license  GPL-2.0+
+ */
 
 if ( ! defined('ABSPATH')) {
-	exit;
+	exit; // Exit if accessed directly.
 }
-// Exit if accessed directly
 
 if ( ! function_exists('auto_load_next_post_is_ajax')) {
 	/**
@@ -37,14 +37,15 @@ if ( ! function_exists('supports_alnp')) {
 	/**
 	 * Returns true or false if the plugin is supported by the theme.
 	 *
-	 * @since  1.4.3
-	 * @access public
-	 * @return boolen
+	 * @since   1.4.3
+	 * @version 1.4.8
+	 * @access  public
+	 * @return  boolen
 	 */
 	function supports_alnp() {
 		/* WordPress core themes. */
 		$core_themes = array(
-			'twentyfifteen', 'twentyfourteen', 'twentythirteen', 'twentytwelve', 'twentyten'
+			'twentyseventeen', 'twentyfifteen', 'twentyfourteen', 'twentythirteen', 'twentytwelve', 'twentyten'
 		);
 
 		if (in_array(get_option('template'), $core_themes)) {
@@ -55,4 +56,31 @@ if ( ! function_exists('supports_alnp')) {
 
 		return false;
 	} // END supports_alnp()
+}
+
+if ( ! function_exists('alnp_template_location')) {
+	/**
+	 * Filters the template location for get_template_part().
+	 *
+	 * @since   1.4.8
+	 * @access  public
+	 * @return  boolen
+	 */
+	function alnp_template_location() {
+		$current_theme = get_option('template');
+
+		switch( $current_theme ) {
+			case 'twentyseventeen':
+				$path = 'template-parts/post/';
+				break;
+
+			default:
+				$path = '';
+				break;
+		}
+
+		return $path;
+	} // END alnp_template_location()
+
+	add_filter('alnp_template_location', 'alnp_template_location');
 }